final receptor.py

+++ receptor.py
@@ -1,8 +1,11 @@
+import os
 import cv2
 import numpy as np
 import time
 
 # Função para detectar a cor (preto ou branco)
+
+
 def detectar_cor(frame):
     # Definir área central
     altura, largura, _ = frame.shape
@@ -60,18 +63,21 @@ def detectar_cor(frame):
     else:
         return "Preto"
 
+
 contBit = 0
 # Função para registrar a cor detectada em um arquivo .txt
+
+
 def registrar_cor(cor):
-    with open("registro_cores.txt", "a", encoding='utf-8') as arquivo:
+    with open("registroBinario.txt", "a", encoding='utf-8') as arquivo:
         if cor == "Preto":
             global contBit
-            contBit+=1
+            contBit += 1
             print(contBit)
             arquivo.write("1")
 
         elif cor == "Branco":
-            contBit+=1
+            contBit += 1
             print(contBit)
             arquivo.write("0")
 
@@ -128,14 +134,13 @@ while True:
     if registro_ativo and not processo_finalizado:
         if cor in ["Preto", "Branco"] and delayContagem:
             print("entrou")
-            time.sleep(0.12) #0.12 = 14dig. em=1000
+            time.sleep(0.12)  # 0.12 = 14dig. em=1000
             delayContagem = False
-        
+
         # Calcular o tempo restante para o próximo registro
         tempo_atual = time.time()
         tempo_restante = max(
             0, int(intervalo_tempo - (tempo_atual - ultimo_registro)))
-        
 
         if cor in ["Preto", "Branco"] and tempo_restante == 0:
             # Registrar a cor detectada
@@ -161,18 +166,21 @@ while True:
 
     # Sair ao pressionar a tecla 'q'
     if cv2.waitKey(1) & 0xFF == ord('q'):
-        with open("registroBinario.txt", "r", encoding='utf-8') as file:
-            binary_data = file.read()
+        if os.path.exists("registroBinario.txt"):  # Verifica se o arquivo existe
+            with open("registroBinario.txt", "r", encoding='utf-8') as file:
+                binary_data = file.read()
 
-        if len(binary_data) % 8 != 0:
-            print("Erro: A sequência binária não tem múltiplos de 8 bits!")
+            if len(binary_data) % 8 != 0:
+                print("Erro: A sequência binária não tem múltiplos de 8 bits!")
+            else:
+                converted_text = binary_to_text(binary_data)
+
+                with open("texto_traduzido.txt", "w", encoding='utf-8') as output_file:
+                    output_file.write(converted_text)
+
+                print("Texto convertido e salvo em 'texto_traduzido.txt'.")
         else:
-            converted_text = binary_to_text(binary_data)
-
-            with open("texto_traduzido.txt", "w", encoding='utf-8') as output_file:
-                output_file.write(converted_text)
-
-            print("Texto convertido e salvo em 'texto_traduzido.txt'.")
+            print("O arquivo 'registroBinario.txt' não foi encontrado.")
         break
 
 # Liberar a captura e fechar as janelas
